requirements_pure_llm.txt
# Pure LLM Architecture Requirements
litellm>=1.0.0
python-dotenv>=1.0.0